fastboot.sh: reverting the patch #693.
[Comment]
This fastboot.sh should be run from the /out/target/product/blaze_tablet/
directory and not from the device/ti/blaze_tablet/ directory.
During build, the binaries and this fastboot.sh will be copied to the out directory.
So the path for the binaries mentioned in the fastboot.sh was correct.

[DR]
-

Change-Id: I84140cd8622b77002b0400cd8a074abc7411bbe9
Signed-off-by: Solaiyappan Saravanan <saravanan.s@ti.com>