Changed Ducati binaries copy method.
Changed the build so that Blaze Ducati binaries are included in
the local builds as well as the Daily builds.  Added the Blaze
Ducati binaries to vendor/ti/blaze in
https://android.dal.design.ti.com/#change,3797.  Now changed
the device.mk file to use the device-vendor.mk file from there.

Equivalent change to the device/ti/blaze patch:
https://android.dal.design.ti.com/#change,3798

Change-Id: If93ff38ac4e7134a0f6510d9ac63d71136b030ac
Signed-off-by: Jean Johnson <a0271255@ti.com>