[Tablet] Enable nomblk_io_submit mount option for read-write ext4 filesystems.
The new mblk_io_submit code was turned on by default in 2.6.39, but it seems
have a bug that we hit on stingray.  The new code was to improve performance
on 48-way systems, which we don't need on Android, so turn it off and use the
previous method by mounting with nomblk_io_submit.

Change-Id: I1250878d896aeab0e3c7102b9f7c9bfc46e26345
Signed-off-by: Dan Murphy <dmurphy@ti.com>