crypto: put android crypto footer on new crypto partition
This places the crypto footer used by Android's crypto implementation
into a new partition "crypto" created to for this purpose.
This footer needs to be in unformatted storage that is read/write.

The Android crypto footer contains details on the type of encryption, and an
encrypted copy of the master key to decrypt the filesystem. Android uses
this footer without a filesystem, directly accessing the block device.

Change-Id: I3f056b1bfca31a5b346e1ed793e16cd7091ab2ec
Signed-off-by: Bryan Buckley <bryan.buckley@ti.com>