init.omap4blazeboard.rc: Convert services to class main
Encryption was failing because some services are not allowing /data
to be unmounted.

core services are never shut down after starting
main services are shut down and restarted after the framework is restarted
late_start services are not started until after /data has been decrypted
  and mounted

Thus if a service is a core service and is using a file on /data,
unmounting data will fail while trying encryption.

Change-Id: I9f86d873b990ac20ed2473694c1592fbfe937d8b
Signed-off-by: Bryan Buckley <bryan.buckley@ti.com>