blaze_tablet: fix fastboot assumption on relative directory
Rather than use a relative directory, use Android build variables
to pick up the fastboot binary location and location of the AFS
images.

Also pick up on whether we're in a Android FS and assume
that we're in a daily build tar ball and use the current directory
as the location for fastboot and the location of the AFS images.

Change-Id: I8b5a4d528d7b1394f77aea3415db1d5cc22602cb
Signed-off-by: Tony Lofthouse <a0741364@ti.com>