blaze_tablet: lights: avoid double setting same value
Avoid setting same value of brightness twice in a row.
This is a T2 specific issue. If we set brightness to 0 twice in a row
then LCD is disable too. Brightness animation that comes from userspace
use zero value as end of fade animation as well as start value of rising
up one. So actually rasing up animation disable LCD and then increase
brightness of LED. Fix this.

Change-Id: Id47e6d8b1eb4f416ea430e12b1eec30fe3c1bef2
Signed-off-by: Andrii Beregovenko <a.beregovenko@ti.com>