init.rc: changes for HSIC support on Tablet/Tablet2
In case of replacing USB2640 by USB4640 (HSIC hub)
on Tablet/Tablet2 it is connected to the second port
of the root hub, so all the changes for the wakeup
enabling etc should be added to that port.
This will work for the Hub that is really present
in the board. If the Hub is absent, then these lines
will return invisible for user errors.

Change-Id: I28612157b789867a1499cfdeedacfb78a4b45f95
Signed-off-by: Ruslan Bilovol <ruslan.bilovol@ti.com>