Blaze_Tablet: Added System Property to know the active Endpoint
The DD+ decoder relies upon a system property "dolby.audio.sink.info"
to give information about the connected endpoint and capability.
The decoder changes its output number of channels and the Output Reference
Levels based on the value of this property. This property allows distinction
between HDMI 6-channel, 2-channel and other endpoints.

By default this property is set to SPEAKER.

Change-Id: Ib97211e194691bf760bd444ea47b0d4906c8c383
Signed-off-by: Shathiesh Rao <shathiesh.rao@ti.com>