blaze_tablet: Enable BlueTI
BlueTI is the Texas Instruments's Bluetooth stack based on BlueZ. The BlueZ
version of a particular Android pastry is upgraded and extra features are
added on top of it.

This requires the Bluetooth sub-system and the Bluetooth drivers to be built
and inserted on boot as modules via the compat-wireless approach.
Flags to enable BlueTI enhancement are both compiled time and runtime system
property.

Change-Id: If0cf703dd61571157d6fe31c26b48e844a35fc58
Signed-off-by: Pavan Savoy <pavan_savoy@ti.com>