blaze_tablet: enable screen wakeup on charger/usb plug/unplug
In previous android releases, the screen was waking up on
charger/usb plug/unplug events.
In latest jellybean sources google became more flexible and
added ability to change this behavior. By default this feature
was disabled and we need to enable it manually,

This patch enables it using android overlay technique like
it has been done for other GEDs.

Change-Id: I3a598097e3ca54dfa741ca8aed363d3a6c8cd843
Signed-off-by: Ruslan Bilovol <ruslan.bilovol@ti.com>